# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Thanks to Cerebras API",
    "version": "1.0.0",
    "description": "OpenAPI schema for the OpenAI-compatible proxy and same-origin admin API. Every response includes x-request-id; clients may send X-Request-Id to propagate a correlation ID."
  },
  "servers": [{ "url": "/" }],
...
  "paths": {
...
    "/api/keys": {
      "get": {
        "tags": ["Admin keys"],
        "operationId": "listApiKeys",
        "summary": "List stored Cerebras API key metadata",
        "security": [{ "AdminToken": [] }],
        "responses": {
          "200": {
            "description": "API key metadata list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiKeyList" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Problem" }
        }
      },
      "post": {
        "tags": ["Admin keys"],
        "operationId": "createApiKey",
        "summary": "Add a Cerebras API key",
        "security": [{ "AdminToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateApiKeyRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateApiKeyResponse"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Problem" },
          "401": { "$ref": "#/components/responses/Problem" },
          "409": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/api/keys/batch": {
      "post": {
        "tags": ["Admin keys"],
        "operationId": "batchImportApiKeys",
        "summary": "Import multiple Cerebras API keys",
        "security": [{ "AdminToken": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BatchImportRequest" }
            },
            "text/plain": { "schema": { "type": "string" } }
          }
        },
        "responses": {
          "200": {
            "description": "Batch import summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BatchImportResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/Problem" },
          "401": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/api/keys/export": {
      "get": {
        "tags": ["Admin keys"],
        "operationId": "exportAllApiKeys",
        "summary": "Reject plaintext API key export",
        "security": [{ "AdminToken": [] }],
        "responses": { "403": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/api/keys/{id}": {
      "delete": {
        "tags": ["Admin keys"],
        "operationId": "deleteApiKey",
        "summary": "Delete a Cerebras API key",
        "security": [{ "AdminToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Success" },
          "401": { "$ref": "#/components/responses/Problem" },
          "404": { "$ref": "#/components/responses/Problem" },
          "409": { "$ref": "#/components/responses/Problem" },
          "500": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
    "/api/keys/{id}/export": {
      "get": {
        "tags": ["Admin keys"],
        "operationId": "exportApiKey",
        "summary": "Reject plaintext API key export",
        "security": [{ "AdminToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": { "403": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/api/keys/{id}/test": {
      "post": {
        "tags": ["Admin keys"],
        "operationId": "testApiKey",
        "summary": "Test a Cerebras API key against the current model pool",
        "security": [{ "AdminToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/IdPath" }],
        "responses": {
          "200": {
            "description": "API key test result",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TestResult" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Problem" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      },
      "SetupToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Setup-Token"
      },
      "ProxyBearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Proxy key supplied as Authorization: Bearer <proxy-key>. May be bypassed when proxyPublicAccess is enabled."
      }
    },
    "headers": {
      "XRequestId": {
        "description": "Correlation ID echoed from X-Request-Id or generated per request",
        "schema": { "type": "string" }
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string" }
      },
...
    },
...
    "responses": {
      "Problem": {
        "description": "Problem Details error",
        "headers": {
          "x-request-id": { "$ref": "#/components/headers/XRequestId" }
        },
        "content": {
          "application/problem+json": {
            "schema": { "$ref": "#/components/schemas/Problem" }
          }
        }
      },
...
      "Success": {
        "description": "Successful operation",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Success" }
          }
        }
      },
...
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "required": ["type", "title", "status", "detail"],
        "properties": {
          "type": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "detail": { "type": "string" },
          "instance": { "type": "string" }
        },
        "additionalProperties": false
      },
...
      "Success": {
        "type": "object",
        "required": ["success"],
        "properties": { "success": { "type": "boolean", "const": true } },
        "additionalProperties": true
      },
...
      "ApiKeyList": {
        "type": "object",
        "required": ["keys"],
        "properties": {
          "keys": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ApiKeyMetadata" }
          }
        },
        "additionalProperties": false
      },
      "ApiKeyMetadata": {
        "type": "object",
        "required": ["id", "useCount", "status", "createdAt"],
        "properties": {
          "id": { "type": "string" },
          "useCount": { "type": "integer", "minimum": 0 },
          "lastUsed": { "type": "integer" },
          "status": {
            "type": "string",
            "enum": ["active", "inactive", "invalid"]
          },
          "createdAt": { "type": "integer" }
        },
        "additionalProperties": false
      },
      "CreateApiKeyRequest": {
        "type": "object",
        "required": ["key"],
        "properties": { "key": { "type": "string" } },
        "additionalProperties": false
      },
      "CreateApiKeyResponse": {
        "type": "object",
        "required": ["success", "id"],
        "properties": {
          "success": { "type": "boolean", "const": true },
          "id": { "type": "string" }
        },
        "additionalProperties": true
      },
      "BatchImportRequest": {
        "oneOf": [
          {
            "type": "object",
            "properties": { "input": { "type": "string" } },
            "additionalProperties": true
          },
          { "type": "string" }
        ]
      },
      "BatchImportResponse": {
        "type": "object",
        "required": ["summary", "results"],
        "properties": {
          "summary": {
            "type": "object",
            "required": ["total", "success", "failed"],
            "properties": {
              "total": { "type": "integer" },
              "success": { "type": "integer" },
              "failed": { "type": "integer" }
            },
            "additionalProperties": false
          },
          "results": {
            "type": "object",
            "required": ["success", "failed"],
            "properties": {
              "success": {
                "type": "array",
                "items": { "type": "string" }
              },
              "failed": {
                "type": "array",
                "items": {
                  "type": "object",
                  "required": ["key", "error"],
                  "properties": {
                    "key": { "type": "string" },
                    "error": { "type": "string" }
                  },
                  "additionalProperties": false
                }
              }
            },
            "additionalProperties": false
          }
        },
        "additionalProperties": false
      },
...
      "TestResult": {
        "type": "object",
        "required": ["success", "status"],
        "properties": {
          "success": { "type": "boolean" },
          "status": { "type": "string" },
          "error": { "type": "string" }
        },
        "additionalProperties": false
      },
...
    }
  }
}